+++ redis/handler.go
@@ -18,10 +18,12 @@ package redis
 type CommandHandler interface {
 	Set(ctx *DBContext, key string, val string, opt SetOption) (*Message, error)
 	Get(ctx *DBContext, key string, opt GetOption) (*Message, error)
-	HSet(ctx *DBContext, hash string, key string, val string, opt HSetOption) (*Message, error)
-	HGet(ctx *DBContext, hash string, key string, opt HGetOption) (*Message, error)
 	MSet(ctx *DBContext, dict map[string]string, opt MSetOption) (*Message, error)
 	MGet(ctx *DBContext, keys []string, opt MGetOption) (*Message, error)
+	HSet(ctx *DBContext, hash string, key string, val string, opt HSetOption) (*Message, error)
+	HGet(ctx *DBContext, hash string, key string, opt HGetOption) (*Message, error)
+	HMSet(ctx *DBContext, hash string, dict map[string]string, opt HMSetOption) (*Message, error)
+	HMGet(ctx *DBContext, hash string, keys []string, opt HMGetOption) (*Message, error)
 }
 
 // SystemCommandHandler is a hander interface for system commands.